+++ demo/src/.vuepress/config.ts
@@ -54,7 +54,7 @@ export default hope.config({
       },
       {
         text: "Docs",
-        link: "https://vuepress-theme-hope.github.io/",
+        link: "https://vuepress-theme-hope.github.io/v1/",
         icon: "note",
       },
     ],
@@ -87,7 +87,7 @@ export default hope.config({
           {
             text: "主题文档",
             icon: "note",
-            link: "https://vuepress-theme-hope.github.io/zh/",
+            link: "https://vuepress-theme-hope.github.io/v1/zh/",
           },
         ],
         sidebar: {