+++ docs/theme/src/.vuepress/config.ts
@@ -6,6 +6,8 @@ export default hope.config({
   title: "vuepress-theme-hope",
   description: "A vuepress theme with tons of features✨",
 
+  base: "/v1/",
+
   dest: "./dist",
 
   head: [