+++ packages/add-this/src/types/index.d.ts
@@ -3,11 +3,11 @@ export interface AddThisOptions {
   /**
    * Public ID for add this
    *
-   * @see https://vuepress-theme-hope.github.io/add-this/guide/
+   * @see https://vuepress-theme-hope.github.io/v1/add-this/guide/
    *
    * Add This 的公开 ID
    *
-   * @see https://vuepress-theme-hope.github.io/add-this/zh/guide/
+   * @see https://vuepress-theme-hope.github.io/v1/add-this/zh/guide/
    */
   pubid: string;
 }